backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
httpx
openai
sqlalchemy
psycopg[binary]
chromadb
python-multipart
pypdf
python-docx
pandas
openpyxl
lxml
xmlschema
